requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
frozenlist==1.3.3
idna==3.4
iniconfig==2.0.0
lxml==4.9.2
multidict==6.0.4
openai==0.27.4
packaging==23.1
pluggy==1.0.0
pydantic==1.10.7
pytest==7.3.1
python-dotenv==1.0.0
requests==2.28.2
rss-parser==0.2.4
soupsieve==2.4.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.15
yarl==1.8.2
